+++ R/Hancock-package.R
@@ -1,6 +1,7 @@
 #' @keywords internal
 #' @import SummarizedExperiment
 #' @import GSEABase
+#' @import GeneSet
 #' @importFrom S4Vectors metadata metadata<-
 #' @importFrom methods setGeneric getPackageName
 #' @importFrom utils packageVersion