+++ tests/testthat/setup-library.R
@@ -1,1 +1,2 @@
 library(SummarizedExperiment)
+library(GeneSet)