+++ tests/testthat/test-GeneSet.R
@@ -1,15 +1,15 @@
 
 # Example data ----
 
-gsc <- GeneSetCollection(list(
-    GeneSet(setName="Cell type 1", c("Gene001", "Gene002")),
-    GeneSet(setName="Cell type 2", c("Gene002", "Gene003", "Gene004"))
-))
+tgs <- tbl_geneset(
+    "Cell type 1" = c("Gene001", "Gene002"),
+    "Cell type 2" = c("Gene002", "Gene003", "Gene004")
+)
 
 # .makeFilterExpressionFromGeneSetCollection ----
 
 test_that(".makeFilterExpressionFromGeneSetCollection works", {
-    out <- makeFilterExpression(gsc)
+    out <- makeFilterExpression(tgs)
 
     expect_named(out, c("Cell type 1", "Cell type 2"))
 })
@@ -17,6 +17,6 @@ test_that(".makeFilterExpressionFromGeneSetCollection works", {
 # .uniqueMarkers ----
 
 test_that(".uniqueMarkers works", {
-    out <- uniqueMarkers(gsc)
+    out <- uniqueMarkers(tgs)
     expect_identical(out, c("Gene001", "Gene002", "Gene003", "Gene004"))
 })